show friendly error on missing file in actions
instead of an unhelpful 404 error show a message and refresh button.

+++ src/files/views.py
@@ -1,8 +1,17 @@
+import contextlib
+
+from django.core.exceptions import ObjectDoesNotExist
 from django.http import HttpResponse
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import render
 
 from files.forms import FileForm
 from files.models import File
+
+
+def get_object_or_gone(request, pk: int):
+    with contextlib.suppress(ObjectDoesNotExist):
+        return File.objects.get(pk=pk), None
+    return None, render(request, "files/file-not-found.html")
 
 
 def file_list_results(request):
@@ -11,17 +20,23 @@ def file_list_results(request):
 
 
 def file_actions(request, pk: int):
-    file = get_object_or_404(File, pk=pk)
+    file, err = get_object_or_gone(request, pk=pk)
+    if err:
+        return err
     return render(request, "files/file-actions.html", {"file": file})
 
 
 def file_actions_trigger(request, pk: int):
-    file = get_object_or_404(File, pk=pk)
+    file, err = get_object_or_gone(request, pk=pk)
+    if err:
+        return err
     return render(request, "files/file-actions-trigger.html", {"file": file})
 
 
 def file_action_delete(request, pk: int):
-    file = get_object_or_404(File, pk=pk)
+    file, err = get_object_or_gone(request, pk=pk)
+    if err:
+        return err
     if "confirm" in request.POST:
         file.delete()
         return HttpResponse(headers={"HX-Trigger-After-Settle": "fileList"})
